+++ src/Koopa.h
@@ -18,7 +18,7 @@
 class Koopa : public GameObject
 {
 	virtual void GetBoundingBox(float& left, float& top, float& right, float& bottom);
-	virtual void Update(DWORD dt, std::vector<LPGAMEOBJECT>* coObjects);
+	virtual void Update(ULONGLONG dt, std::vector<LPGAMEOBJECT>* coObjects);
 	virtual void Render();
 
 public:
